requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
google-api-python-client>=2.136.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-generativeai>=0.7.2
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.4
requests>=2.32.3
gunicorn
uvloop
httptools
